fix(agent): keep server routes lightweight
