+++ packages/agent/test/config.test.ts
@@ -77,14 +77,22 @@ describe("agent config", () => {
     }
   })
 
-  it("exposes channel route helpers from the Agent server import", async () => {
+  it("keeps workspace runtime loading out of the Agent server route import", async () => {
     vi.resetModules()
+    vi.doMock("@vite-hub/workspace", () => {
+      throw new Error("workspace should only load when a workspace-backed path is used")
+    })
+    vi.doMock("@vite-hub/workspace/runtime", () => {
+      throw new Error("workspace runtime should only load when workspace registration is used")
+    })
     try {
       const server = await import("../src/server.ts")
       expect(server.createChannelChatRouteHandler).toBeTypeOf("function")
       expect(server.createChannelWebhookRouteHandler).toBeTypeOf("function")
     }
     finally {
+      vi.doUnmock("@vite-hub/workspace")
+      vi.doUnmock("@vite-hub/workspace/runtime")
       vi.resetModules()
     }
   })
